Add jupytext use via pre-commit and pair the notebook files (#3)
* Add jupytext configuration

* Remove old .py files

* Add paired .py notebooks

* [pre-commit.ci] auto fixes from pre-commit.com hooks



jupytext provides tools to pair .ipynb-format notebooks and .py-format notebooks (which work more easily in non-jupyter tools and provide clearer git diffs). It's useful to use pre-commit to run jupytext (configured via the .pre-commit-config.yaml file).

There's a few different standards for the .py-format notebook. I'm using the most common percent format. It will work with IDEs like VSCode and Spyder. The main difference with the ipynb format is that it does not contain the cell outputs.

pre-commit can be run before committing locally, but it's not necessary when we add it to the PR workflow, which I have done by installing the https://pre-commit.ci/ app to the repo. The app will run pre-commit on each PR and push an extra commit itself if needed to synchronise the .ipynb/.py files. You can see some fail examples below before pre-commit.ci pushes the fix that makes it pass.

All this means that as a developer, you edit either the .py file or the .ipynb but not both, then the automation handles the synchronisation of the two during the PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>